Split view in seperate view tabs
